ir: fix clang-tidy warnings

+++ include/bal_types.h
@@ -10,7 +10,7 @@
 
 
 typedef uint64_t bal_guest_address_t;
-typedef uint32_t bal_instruction_t;
+typedef uint64_t bal_instruction_t;
 typedef uint32_t bal_constant_t;
 typedef uint16_t bal_instruction_count_t;
 typedef uint16_t bal_ssa_id_t;
